requirements.txt
Flask==2.2.3
pyserial==3.5
